Moving more math over to maths module
